requirements.txt
# AutoDBAudit - Production Dependencies
# Minimal, production-ready dependencies for PyInstaller compatibility

# SQL Server connectivity (ODBC-based, SQL 2008 R2 through 2022+ compatible)
pyodbc>=5.0.0,<6.0.0

# Excel generation (pure Python, no Excel installation required)
# Supports advanced formatting: icons, charts, conditional formatting
openpyxl>=3.1.0,<4.0.0

# Rich terminal output (colorful CLI help and progress)
rich>=13.0.0

# Image support for Excel (required by openpyxl for inserting images)
pillow>=10.0.0

# Windows API access for DPAPI credential encryption
# Platform-specific: Windows only
pywin32>=306
